requirements.txt
langchain
langchain-google-genai
python-dotenv
google-generativeai>=0.8.3
chromadb
pydantic
langchain-chroma
langgraph
